Prototype having the whiteboard launch implementations as separate bundles
