+++ org.gecko.rest.jersey.jetty/src/org/gecko/rest/jersey/jetty/JettyBackedWhiteboardComponent.java
@@ -11,19 +11,19 @@
  *     Stefan Bishof - API and implementation
  *     Tim Ward - implementation
  */
-package org.gecko.rest.jersey.runtime.httpwhiteboard;
+package org.gecko.rest.jersey.jetty;
 
+import static org.osgi.service.component.annotations.ReferenceCardinality.AT_LEAST_ONE;
 import static org.osgi.service.component.annotations.ReferenceCardinality.MULTIPLE;
 import static org.osgi.service.component.annotations.ReferencePolicy.DYNAMIC;
+import static org.osgi.service.jakartars.whiteboard.JakartarsWhiteboardConstants.JAKARTA_RS_EXTENSION;
 import static org.osgi.service.jakartars.whiteboard.JakartarsWhiteboardConstants.JAKARTA_RS_RESOURCE;
 
 import java.util.Map;
 import java.util.logging.Logger;
 
-import jakarta.ws.rs.core.Application;
-
 import org.gecko.rest.jersey.provider.JerseyConstants;
-import org.gecko.rest.jersey.runtime.JerseyWhiteboardComponent;
+import org.gecko.rest.jersey.runtime.AbstractWhiteboard;
 import org.osgi.framework.ServiceObjects;
 import org.osgi.framework.ServiceReference;
 import org.osgi.service.cm.ConfigurationException;
@@ -35,39 +35,42 @@ import org.osgi.service.component.annotations.ConfigurationPolicy;
 import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Modified;
 import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ReferenceCardinality;
-import org.osgi.service.component.annotations.ReferencePolicy;
 import org.osgi.service.condition.Condition;
 import org.osgi.service.jakartars.whiteboard.JakartarsWhiteboardConstants;
 
+import jakarta.ws.rs.core.Application;
+
 /**
- * This component handles the lifecycle of a {@link JakartarsServiceRuntime}
+ * A configurable component, that establishes a whiteboard
  * @author Mark Hoffmann
- * @since 30.07.2017
+ * @since 11.10.2017
  */
-@Component(name="JakartarsHttpWhiteboardRuntimeComponent", immediate=true, configurationPolicy=ConfigurationPolicy.REQUIRE, reference = @Reference(name = "runtimeCondition", service = Condition.class , target = JerseyConstants.JERSEY_RUNTIME_CONDITION))
-public class JakartarsHttpWhiteboardRuntimeComponent extends JerseyWhiteboardComponent{
 
-	private static Logger logger = Logger.getLogger("o.e.o.j.JakartarsHttpWhiteboardRuntimeComponent");
+@Component(name = "JakartarsWhiteboardComponent", 
+	reference = @Reference(name = "runtimeCondition", service = Condition.class , target = JerseyConstants.JERSEY_RUNTIME_CONDITION),
+	configurationPolicy = ConfigurationPolicy.REQUIRE
 
+)
+public class JettyBackedWhiteboardComponent extends AbstractWhiteboard {
+
+	Logger logger = Logger.getLogger("o.e.o.j.runtimeComponent");
 
 	/**
 	 * Called on component activation
 	 * @param componentContext the component context
 	 * @throws ConfigurationException 
 	 */
-	/* (non-Javadoc)
-	 * @see org.gecko.rest.jersey.runtime.JerseyWhiteboardComponent#activate(org.osgi.service.component.ComponentContext)
-	 */
 	@Activate
-	@Override
-	public void activate(final ComponentContext componentContext) throws ConfigurationException {
+	public void activate(ComponentContext componentContext) throws ConfigurationException {
 		updateProperties(componentContext);
+		
 		if (whiteboard != null) {
-			whiteboard.teardown();;
+			whiteboard.teardown();
 		}
-		whiteboard = new HTTPWhiteboardBasedJerseyServiceRuntime();
+		whiteboard = new JerseyServiceRuntime();
+		// activate and start server
 		whiteboard.initialize(componentContext);
+//		dispatcher.setBatchMode(true);
 		dispatcher.setWhiteboardProvider(whiteboard);
 		dispatcher.dispatch();
 		whiteboard.startup();
@@ -81,8 +84,8 @@ public class JakartarsHttpWhiteboardRuntimeComponent extends JerseyWhiteboardCom
 	@Modified
 	public void modified(ComponentContext context) throws ConfigurationException {
 		updateProperties(context);
-		whiteboard.modified(context);
 		dispatcher.dispatch();
+		whiteboard.modified(context);
 	}
 
 	/**
@@ -99,14 +102,14 @@ public class JakartarsHttpWhiteboardRuntimeComponent extends JerseyWhiteboardCom
 			whiteboard = null;
 		}
 	}
-
+	
 	/**
 	 * Adds a new default application
 	 * @param application the application to add
 	 * @param properties the service properties
 	 */
-	@Reference(name="defaultApplication", cardinality=ReferenceCardinality.AT_LEAST_ONE, policy=ReferencePolicy.DYNAMIC, unbind="unbindDefaultApplication", updated = "modifedDefaultApplication", target="(&(" + JakartarsWhiteboardConstants.JAKARTA_RS_APPLICATION_BASE	+ "=*)(" + JakartarsWhiteboardConstants.JAKARTA_RS_NAME	+ "=.default))")
-	public void addDefaultApplication(Application application, Map<String, Object> properties) {
+	@Reference(cardinality = AT_LEAST_ONE, policy = DYNAMIC, target = "(&(" + JakartarsWhiteboardConstants.JAKARTA_RS_APPLICATION_BASE + "=*)(" + JakartarsWhiteboardConstants.JAKARTA_RS_NAME + "=.default))")
+	public void bindDefaultApplication(Application application, Map<String, Object> properties) {
 		dispatcher.addApplication(application, properties);
 	}
 
@@ -128,17 +131,18 @@ public class JakartarsHttpWhiteboardRuntimeComponent extends JerseyWhiteboardCom
 	public void unbindDefaultApplication(Application application, Map<String, Object> properties) {
 		dispatcher.removeApplication(application, properties);
 	}
-
+	
 	/**
 	 * Adds a new application
 	 * @param application the application to add
 	 * @param properties the service properties
 	 */
-	@Reference(name="application", service=Application.class,cardinality=ReferenceCardinality.MULTIPLE, policy=ReferencePolicy.DYNAMIC, unbind="unbindApplication", updated = "modifedApplication", target="(&(" + JakartarsWhiteboardConstants.JAKARTA_RS_APPLICATION_BASE	+ "=*)(!(" + JakartarsWhiteboardConstants.JAKARTA_RS_NAME	+ "=.default)))")
+	@Reference(service = Application.class, cardinality = MULTIPLE, policy = DYNAMIC, target = "(&(" + JakartarsWhiteboardConstants.JAKARTA_RS_APPLICATION_BASE + "=*)(!(" + JakartarsWhiteboardConstants.JAKARTA_RS_NAME + "=.default)))")
 	public void bindApplication(Application application, Map<String, Object> properties) {
 		dispatcher.addApplication(application, properties);
+	
 	}
-
+	
 	/**
 	 * Adds a new application
 	 * @param application the application to add
@@ -148,7 +152,7 @@ public class JakartarsHttpWhiteboardRuntimeComponent extends JerseyWhiteboardCom
 		dispatcher.removeApplication(application, properties);
 		dispatcher.addApplication(application, properties);
 	}
-
+	
 	/**
 	 * Removes a application 
 	 * @param application the application to remove
@@ -158,24 +162,28 @@ public class JakartarsHttpWhiteboardRuntimeComponent extends JerseyWhiteboardCom
 		dispatcher.removeApplication(application, properties);
 	}
 
-	@Reference(service = AnyService.class, target = "(" + JakartarsWhiteboardConstants.JAKARTA_RS_EXTENSION	+ "=true)", cardinality = MULTIPLE, policy = DYNAMIC)
+	@Reference(service = AnyService.class, target = "(" + JAKARTA_RS_EXTENSION
+			+ "=true)", cardinality = MULTIPLE, policy = DYNAMIC)
 	public void bindJakartarsExtension(ServiceReference<Object> jakartarsExtensionSR, Map<String, Object> properties) {
-		unbindJakartarsExtension(jakartarsExtensionSR,properties);
+
+		updatedJakartarExtension(jakartarsExtensionSR, properties);
 	}
 
-	public void updatedJakartarsExtension(ServiceReference<Object> jakartarsExtensionSR, Map<String, Object> properties) {
+	public void updatedJakartarExtension(ServiceReference<Object> jakartarsExtensionSR, Map<String, Object> properties) {
 		logger.fine("Handle extension " + jakartarsExtensionSR + " properties: " + properties);
 		ServiceObjects<?> so = getServiceObjects(jakartarsExtensionSR);
 		dispatcher.addExtension(so, properties);
 
 	}
+
 	public void unbindJakartarsExtension(ServiceReference<Object> jakartarsExtensionSR, Map<String, Object> properties) {
 		dispatcher.removeExtension(properties);
 	}
 
-	@Reference(service = AnyService.class, target = "(" + JAKARTA_RS_RESOURCE + "=true)", cardinality = MULTIPLE, policy = DYNAMIC)
+	@Reference(service = AnyService.class, target = "(" + JAKARTA_RS_RESOURCE
+			+ "=true)", cardinality = MULTIPLE, policy = DYNAMIC)
 	public void bindJakartarsResource(ServiceReference<Object> jakartarsExtensionSR, Map<String, Object> properties) {
-		updatedJakartarsResource(jakartarsExtensionSR,properties);
+		updatedJakartarsResource(jakartarsExtensionSR, properties);
 	}
 
 	public void updatedJakartarsResource(ServiceReference<Object> jakartarsResourceSR, Map<String, Object> properties) {
@@ -184,6 +192,7 @@ public class JakartarsHttpWhiteboardRuntimeComponent extends JerseyWhiteboardCom
 		dispatcher.addResource(so, properties);
 
 	}
+
 	public void unbindJakartarsResource(ServiceReference<Object> jakartarsResourceSR, Map<String, Object> properties) {
 		dispatcher.removeResource(properties);
 	}
