+++ org.gecko.rest.jersey.servlet.whiteboard/src/org/gecko/rest/jersey/runtime/httpwhiteboard/HTTPWhiteboardBasedJerseyServiceRuntime.java
@@ -28,9 +28,9 @@ import jakarta.servlet.Servlet;
 import org.gecko.rest.jersey.helper.JerseyHelper;
 import org.gecko.rest.jersey.provider.JerseyConstants;
 import org.gecko.rest.jersey.provider.application.JakartarsApplicationProvider;
-import org.gecko.rest.jersey.runtime.common.AbstractJerseyServiceRuntime;
-import org.gecko.rest.jersey.runtime.common.ResourceConfigWrapper;
-import org.gecko.rest.jersey.runtime.servlet.WhiteboardServletContainer;
+import org.gecko.rest.jersey.runtime.AbstractJerseyServiceRuntime;
+import org.gecko.rest.jersey.runtime.ResourceConfigWrapper;
+import org.gecko.rest.jersey.runtime.WhiteboardServletContainer;
 import org.glassfish.jersey.server.ResourceConfig;
 import org.glassfish.jersey.servlet.ServletContainer;
 import org.osgi.framework.Bundle;
