+++ org.gecko.rest.jersey/src/org/gecko/rest/jersey/provider/application/package-info.java
@@ -1,1 +1,4 @@
+@Export
 package org.gecko.rest.jersey.provider.application;
+
+import org.osgi.annotation.bundle.Export;
