+++ org.gecko.rest.jersey/src/org/gecko/rest/jersey/provider/whiteboard/package-info.java
@@ -1,1 +1,4 @@
+@Export
 package org.gecko.rest.jersey.provider.whiteboard;
+
+import org.osgi.annotation.bundle.Export;
