+++ org.gecko.rest.jersey/src/org/gecko/rest/jersey/runtime/package-info.java
@@ -1,0 +1,2 @@
+@org.osgi.annotation.bundle.Export
+package org.gecko.rest.jersey.runtime;
