+++ org.gecko.rest.jersey/test/org/gecko/rest/jersey/JakartarsWhiteboardDispatcherTest.java
@@ -32,8 +32,8 @@ import org.gecko.rest.jersey.provider.application.JakartarsWhiteboardDispatcher;
 import org.gecko.rest.jersey.provider.whiteboard.JakartarsWhiteboardProvider;
 import org.gecko.rest.jersey.resources.TestLegacyApplication;
 import org.gecko.rest.jersey.resources.TestResource;
+import org.gecko.rest.jersey.runtime.JerseyWhiteboardDispatcher;
 import org.gecko.rest.jersey.runtime.common.DefaultApplication;
-import org.gecko.rest.jersey.runtime.dispatcher.JerseyWhiteboardDispatcher;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
